Extracting public portion of app controller
