+++ BTCPayServer/Views/Apps/ListApps.cshtml
@@ -53,7 +53,7 @@
                                 {
                                     <a asp-action="@app.UpdateAction" asp-controller="Apps" asp-route-appId="@app.Id">Settings</a><span> - </span>
                                 }
-                                <a asp-action="@app.ViewAction" asp-controller="Apps" asp-route-appId="@app.Id">View</a><span> - </span>
+                                <a asp-action="@app.ViewAction" asp-controller="AppsPublic" asp-route-appId="@app.Id">View</a><span> - </span>
                                 <a asp-action="DeleteApp" asp-route-appId="@app.Id">Remove</a>
                             </td>
                         </tr>
