+++ BTCPayServer/Views/AppsPublic/ViewPointOfSale.cshtml
@@ -1,6 +1,6 @@
 ﻿@inject BTCPayServer.HostedServices.CssThemeManager themeManager
 
-@model ViewPointOfSaleViewModel
+@model BTCPayServer.Models.AppViewModels.ViewPointOfSaleViewModel
 @{
     ViewData["Title"] = Model.Title;
     Layout = null;
